no more warnings in UInfoEx at all
git-svn-id: http://svn.miranda-ng.org/main/trunk@14920 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/UserInfoEx/src/ctrl_contact.cpp
@@ -303,7 +303,7 @@ INT_PTR CALLBACK DlgProc_Phone(HWND hDlg, UINT msg, WPARAM wParam, LPARAM lParam
 					GetDlgItemText(hDlg, EDIT_CATEGORY, cbi->pszCat, cbi->ccCat);
 
 					// save SMS flag
-					if (IsDlgButtonChecked(hDlg, CHECK_SMS) != ((cbi->wFlags & CBEXIF_SMS) == CBEXIF_SMS))
+					if ((int)IsDlgButtonChecked(hDlg, CHECK_SMS) != ((cbi->wFlags & CBEXIF_SMS) == CBEXIF_SMS))
 						cbi->wFlags ^= CBEXIF_SMS;
 				}
 				//fall through
